Move VerifyResponse to Verify namespace
