+++ lib/verify.rb
@@ -18,3 +18,4 @@ end
 
 require_relative "verify/service_provider"
 require_relative "verify/authentication_request"
+require_relative "verify/response"
